backend/requirements.txt
connexion[swagger-ui] >= 2.6.0; python_version>="3.6"
# 2.3 is the last version that supports python 3.4-3.5
connexion[swagger-ui] <= 2.3.0; python_version=="3.5" or python_version=="3.4"
# prevent breaking dependencies from advent of connexion>=3.0
connexion[swagger-ui] <= 2.14.2; python_version>"3.4"
# connexion requires werkzeug but connexion < 2.4.0 does not install werkzeug
# we must peg werkzeug versions below to fix connexion
# https://github.com/zalando/connexion/pull/1044
werkzeug >= 3.0.6; python_version=="3.5" or python_version=="3.4"
swagger-ui-bundle >= 0.0.2
python_dateutil >= 2.6.0
setuptools >= 21.0.0
Flask >= 2.1.1
mysql-connector-python >= 9.1.0
redis >= 5.2.0
pyjwt >= 2.7.0
flask_jwt_extended >= 4.6.0
flask-cors >= 5.0.0
gunicorn >= 23.0.0
rapidfuzz >= 3.11.0
argon2-cffi >= 23.1.0
flask-wtf >= 1.2.2
logging >=0.4.9.6